perf: electron code adjust
