+++ electron/service/logger.ts
@@ -2,7 +2,8 @@ import log4js from "log4js";
 import { once } from "events";
 import { createInterface } from "readline";
 import { createReadStream, statSync } from "fs";
-import { fsCheckFile } from "./file";
+import { join } from "path";
+import { fsCheckFile } from "../utils/file";
 
 export const initLogger = (path: string) => {
   log4js.configure({
@@ -23,12 +24,7 @@ export const initLogger = (path: string) => {
   return logger;
 };
 
-interface LogItem {
-  time: string;
-  type: string;
-  message: string;
-}
-export const readLogByLine = async (path: string, bufferSize = 1024 * 24) => {
+const readLogByLine = async (path: string, bufferSize = 1024 * 24) => {
   await fsCheckFile(path);
   // 默认拿 24KB 数据  应该在300-500条
   const target: LogItem[] = [];
@@ -53,3 +49,9 @@ export const readLogByLine = async (path: string, bufferSize = 1024 * 24) => {
   await once(rl, "close");
   return target;
 };
+
+// 读取最近的几条日志
+export const getRecentLogs = async () => {
+  return readLogByLine(join(process.env.DATA_PATH_LOG, "system.log"));
+};
+
