+++ electron/types/index.d.ts
@@ -1,0 +1,5 @@
+interface LogItem {
+  time: string;
+  type: string;
+  message: string;
+}
